ai-service/pyproject.toml
[project]
name = "web3-talent-ai"
version = "0.1.0"
description = "AI services for Web3 talent matching platform"
authors = [
    {name = "Adam Bryant", email = "adam@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "openai>=1.3.0",
    "PyGithub>=1.59.0",
    "textblob>=0.17.1",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
